% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_track.R
\name{add_track}
\alias{add_track}
\title{Keep track of deleted observations in data cleaning}
\usage{
add_track(data, name, n = nrow(df))
}
\arguments{
\item{data}{This is the dataset df_track, that keeps track of the number of obervations in the main dataset.}

\item{name}{This is a short description of the data restriction}

\item{n}{How many rows are in the main dataset. By default is nrow(df)}
}
\value{
A dataset with an extra line and how many rows df has.
}
\description{
This function allows the user to keep track of the deleted rows
while cleaning the data.
}
\examples{

#df_track <- tibble(
#              Stage = "Original Data",
#              Observations = nrow(df)
#              )
# df_track \%>\% add_track("Delete workers older than 65 years".)


}
