% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_fe.R
\name{D_fe}
\alias{D_fe}
\title{Fitted Fixed effects from fixest}
\usage{
D_fe(data, m)
}
\arguments{
\item{data}{A tibble.}

\item{m}{A fixest object (model).}
}
\value{
A tibble with the estiamted FEs.
}
\description{
This function extracts the FE from a fixest object and then adds it to the data.
Thus, every observations gets their respective estimated FE value for each FE.
The function deals with interactions of FE.
}
\examples{
# df \%<>\% D_fe(m1)


}
