% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_sample.R
\name{event_sample}
\alias{event_sample}
\title{Create a sample with observations that went through an event}
\usage{
event_sample(
  data,
  outcome_var,
  time_var,
  event_var,
  unit,
  change_for_event,
  event_cond = "both",
  stable_n_before,
  stable_n_after,
  stability_cond = "both",
  stability_change,
  did = FALSE,
  ...
)
}
\arguments{
\item{data}{A tibble}

\item{outcome_var}{Not a string. Name of the outcome variable.}

\item{time_var}{Variable with the period indication. Not a string.}

\item{event_var}{Variable representing the event shock. Not a string.}

\item{unit}{Unit of analysis. eg. individual, firm, country. Not a string.}

\item{change_for_event}{Minimum change in event_var necessary to classify as event.}

\item{event_cond}{To be classified as an event, does the change need to be "positive", "negative" or "both"?.}

\item{stable_n_before}{Number of periods before where stability of the event variable must apply}

\item{stable_n_after}{Number of periods after where stability of the event variable must apply}

\item{stability_cond}{Type of stability condition. Can be "event_var", "constant_setting" or "both".}

\item{stability_change}{The maximum change that the outcome variable can have and be called "stable"}

\item{did}{Logical value. Default is FALSE. If true, a control group is returned.}

\item{...}{Stability variables. If "constant_setting" or "both" is selected stability_cond, choose which variables warrant such stability.}
}
\value{
Returns a tidy dataset with observations that classify as an event. If "both" is selected, a column indicating "negative" or positive is also added.
}
\description{
A function to build a sample for the event study.
Consider the event is from t = -1 to t=0
Inspired my Cornelison et al. (2017, AER) event study.
Note: for now, stability_cond = "both" is the only possible condition
The user can join the data with more outcome variables after the sample is created.
}
\examples{
# example code


}
\author{
Miguel Salema
}
