% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_lag.R
\name{fast_lag}
\alias{fast_lag}
\title{Fast lagged/lead values}
\usage{
fast_lag(data, col, group, lag_or_lead, vec_n, new_names, order = year)
}
\arguments{
\item{data}{A tibble.}

\item{col}{Columns to find the lag/lead value. Not a string.}

\item{group}{Group to find lags/leads within. Only one group. Not a string.}

\item{lag_or_lead}{String. Either "lag" or "lead".}

\item{vec_n}{A numeric vector indicating by how many observations should the lag/lead jump.}

\item{new_names}{A character vector with the names of the new variables with the lag/leads.}

\item{order}{Variable to order the dataset by. Not a string. Default is year.}
}
\value{
A tibble with the lagged/leadded columns.
}
\description{
This function creates lagged values faster than dplyr::lag, using data.table.
The function handles multiple lags/leads at once, of the same column.
vec_n and new_names must have the same size.
}
\examples{
# the new names
#names <- paste("lead_wage", 1:5, sep="_")

# run the fast lag/lead function
#df \%<>\% fast_lag(col = wage_lh_real, 
#                 group = worker, 
#                 lag_or_lead = "lead",
#                 vec_n = 1:5,
#                 new_names = names)



}
