% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layers}
\alias{layers}
\title{Layers in QP}
\usage{
layers(data, var_qualif, var_firm, var_year)
}
\arguments{
\item{data}{A tibble}

\item{var_qualif}{A variable with the qualif 1 digit variable. Not a string.}

\item{var_firm}{A variable with the firm unique ID. Not a string.}

\item{var_year}{A variable with the year. Not a string.}
}
\value{
The same dataset with 2 extra columns: layer (the hierarchical position of the worker) and n_layers (the amount of layers in a firm), both from 0 to 3. The first is at the worker level and the second at the firm level.
}
\description{
This function classifies each worker into one of 4 layers.
Furthermore, it counts how many distinct layers exist by firm in each year.

The function first further aggregates the qualif variable and then
produces the layers from it.

This function follows the classification used in the Oppromola paper
However, the only difference is regarding qualif==53 | qualif==55.
Oppromola includes them as layer=0
I follow what Martim, Joana and Jaime did and include it as layer = 1.
This is done due to aggregation limitations, consistent with INE.

In the folder consult_do you can find their stata file.
}
\examples{
# df \%>\%
#   layers(qualif,
#          firm,
#          year)



}
