% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcs.R
\name{lcs}
\alias{lcs}
\title{Largest connected set by two variables}
\usage{
lcs(data, fe1, fe2)
}
\arguments{
\item{data}{A tibble}

\item{fe1}{1st variable}

\item{fe2}{2nd variable}
}
\value{
A dataset with the largest connect set
}
\description{
This function computes the largest connected set for two variables.
In a connected set, all units are connected by atleast one factor in a
transitive manner. That is, I can find a path from one observation to any
other using only the 2 factors provided.
}
\examples{
# Example: AKM
# df \%>\%
#   lcs(firm,
#       firm_lag)



}
