% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_qps.R
\name{load_qps}
\alias{load_qps}
\title{Load QP data}
\usage{
load_qps(name, in_linux = T)
}
\arguments{
\item{name}{Name of the dataset to load. A strting.}

\item{in_linux}{Logical value indicating if the package is being ran on PROSPER's linux server.}
}
\value{
A dataset or a list of available datasets, if "name" is left in blank.
}
\description{
Easily load cleaned QP data.
}
\examples{
df <- load_qps("qp_simple")


}
