% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_event_study.R
\name{plot_event_study}
\alias{plot_event_study}
\title{Ready event data to Plot}
\usage{
plot_event_study(data, outcome_vars, time_var, unit, treatment)
}
\arguments{
\item{data}{A tibble}

\item{outcome_vars}{Character vector with outcome variables.}

\item{time_var}{Variable with the period indication. Treatment occurs from -1 to 0. Not a string. Will be used as a variable in the model.}

\item{unit}{Unit of analysis. eg. individual, firm, country. Not a string. Will be used as a FE}

\item{treatment}{It's the treatment variable.}
}
\value{
A tibble with ready to plot data. The columns are:
}
\description{
The function turns data that comes from an event study or DiD
into a ready to plot dataset of the analysis.
Runs a linear model with FE.
The plot shows confidence intervals.
The data can have been matched.
}
\examples{
# example code
# df_plot_event <- df_event \%>\%
#   plot_event_study(outcome_vars = c("res_wage", "av_peer"),
#                    time_var = t,
#                    unit = worker,
#                    treatment = quality)

}
