% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qp_0.1perc_sample}
\alias{qp_0.1perc_sample}
\title{A 0.1\% sample of QP}
\format{
A data frame with 69466 rows and 68 variables:
\describe{
\item{\code{worker}}{Worker unique ID.}
\item{\code{firm}}{Firm unique ID.}
}
}
\source{
{
Data created in-house to use as an example.
Code:
df <- load_qps("qp+crosswalks+clean+lags+restrictions")
qp_0.1perc_sample <- df \%>\% dplyr::slice_sample(prop = 0.001)
}
}
\usage{
qp_0.1perc_sample
}
\description{
A sample of 0.1 \% of the full cleaned QP with restrictions: "p+crosswalks+clean+lags+restrictions".
The documentation of the variables is not complete.
}
\examples{
data(qp_0.1perc_sample) # lazy loading. The RAM will not be immediately occupied.
}
\keyword{datasets}
