% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_singletons.R
\name{remove_singletons}
\alias{remove_singletons}
\title{Remove Singletons in chosen columns}
\usage{
remove_singletons(data, ..., iter_max = 1)
}
\arguments{
\item{data}{A tibble}

\item{...}{Columns to delete singletons from}

\item{iter_max}{How many iterations to delete singletons. If convergence is achioeved, this is ingnored.}
}
\value{
A dataset with no singletons in the last column chosen.
}
\description{
This function removes singletins in chosen columns.
If more than one column is chosen, it's normal that,
by deleting the singletons in one column, we create singletons in
another columns.
I have a document that indicates that this is not problematic.
The only reason to delete singletons is for computational efficiency.
The estimated coeffficients are not biased and only
in very specific cases is inference harmed.
Thus, the function, by dafault, iterates only one time.
However, the user may choose to do it more times.
}
