% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_measures.R
\name{competition_measures}
\alias{competition_measures}
\title{Worker level competition in QP}
\usage{
competition_measures(data, hiring_var, poach_var, firm_var, wage_var, ...)
}
\arguments{
\item{data}{A tibble}

\item{hiring_var}{Name of the variable indicating if the worker was hired. Not a string.}

\item{poach_var}{Name of the variable indicating id the worker was poached. Not a string.}

\item{firm_var}{The name of the firm identifier, used for HHI. Not a string.}

\item{wage_var}{The name of the variable with total wages. Not a strinf.}

\item{...}{The grouping to compute competitiveness for. Should contain "year".}
}
\description{
This function produces measures of competition using worker level information.
The first measure is "Out of Sample Hires", from Manchin "Monopsony", 1st chapter.
This measures the proportion of workers hired from outside of the sample.
A low proportion of poaches is an indicator of low competition.
The second measure is the Herfindahl Hirschman index (HHI) of employment.
Here, the market share is relative to share of employment and wages,
instead of the typical sales variable.
}
\examples{
# load the simple dataset
data(qp_0.1perc_sample)

qp_0.1perc_sample |>
  competition_measures(hiring,
                       poach,
                       firm,
                       total_wage,
                       year,
                       sector)

}
