% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layers}
\alias{layers}
\title{Layers in QP}
\usage{
layers(data, var_qualif, var_firm, var_year)
}
\arguments{
\item{data}{A tibble}

\item{var_qualif}{A variable with the qualif 1 digit variable. Not a string.}

\item{var_firm}{A variable with the firm unique ID. Not a string.}

\item{var_year}{A variable with the year. Not a string.}
}
\value{
The same dataset with 2 extra columns: layer (the hierarchical position of the worker) and n_layers (the amount of layers in a firm), both from 1 to 4. The first is at the worker level and the second at the firm level.
}
\description{
This function classifies each worker into one of 4 layers,
using the variable "qualif" (professional category),
according to the following table:\tabular{lll}{
   Layer \tab Qualif \tab Classification \cr
   4 \tab 1 \tab Top executives (top management) \cr
   3 \tab 2, 3 \tab Middle management, supervisors, and team leaders \cr
   2 \tab 4, 5 \tab Higher-skilled professionals \cr
   1 \tab 6, 7, 8, 9 \tab Semi- and non-skilled professionals \cr
}


However, the translation from qualif to layer is not immediate.
For the top layer to exist, all other layers must also exist in a firm.

Furthermore, the function also
counts how many distinct layers exist by firm in each year,
from 1 to 4.

I follow the classification of Caliendo et al. (2020),
but I do it from 1 to 4, instead of 0 to 3.
Similarly, I refer to layers and not to layers of management.
However, the only difference is regarding qualif 53 and qualif 55.
Caliendo et al. (2020) includes them as layer=0.
I follow what Martim, Joana and Jaime did and include it as layer = 1.
This is done due to aggregation limitations, consistent with INE.
}
\examples{

data("qp_0.1perc_sample")

# df <- qp_0.1perc_sample \%>\%
#   layers(qualif_1d,
#          firm,
#          year)



}
\references{
Caliendo, L., Mion, G., Opromolla, L. D., & Rossi-Hansberg, E. (2020). Productivity and organization in Portuguese firms. Journal of Political Economy, 128(11), 4211-4257.
}
